added casl for user and roles

+++ libs/auth/src/modules/role/role.controller.ts
@@ -1,62 +1,86 @@
-import { ClassValidationPipe } from '@libs/common/pipes';
+import { Action } from '@libs/auth/constants';
+import { CaslAbility } from '@libs/auth/decorators/current-user';
+import { JwtAuthGuard } from '@libs/auth/guards/jwt-auth-guard';
 import { DurationInterceptor } from '@libs/common/interceptors/duration-interceptor';
+import { ClassValidationPipe } from '@libs/common/pipes';
 import {
-  Controller,
-  Get,
   Body,
+  Controller,
+  Delete,
+  Get,
   Param,
   ParseIntPipe,
-  UseInterceptors,
   Post,
   Put,
-  Delete,
+  UnauthorizedException,
   UseGuards,
+  UseInterceptors,
 } from '@nestjs/common';
+import { CreateRoleInput, UpdateRoleInput } from './dto/role.dto';
 import Role from './role.entity';
 import { RoleService } from './role.service';
-import { CreateRoleInput, UpdateRoleInput } from './dto/role.dto';
-import { Roles } from '@libs/auth/decorators/roles';
-import { JwtAuthGuard } from '@libs/auth/guards/jwt-auth-guard';
-import { RolesGuard } from '@libs/auth/guards/roles-guard';
 
+@UseGuards(JwtAuthGuard)
+@UseInterceptors(DurationInterceptor)
 @Controller('role')
 export class RoleController {
   constructor(private readonly roleService: RoleService) {}
 
-  @Roles('ADMIN')
-  @UseGuards(JwtAuthGuard, RolesGuard)
-  @UseInterceptors(DurationInterceptor)
   @Get()
-  async getRole(): Promise<Role[]> {
-    return this.roleService.findAll();
+  async getRole(@CaslAbility() ability): Promise<Role[]> {
+    if (ability.can(Action.Read, 'role')) {
+      return this.roleService.findAll();
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
-  @UseInterceptors(DurationInterceptor)
   @Get(':id')
-  async findOne(@Param('id', ParseIntPipe) id: number): Promise<Role> {
-    return this.roleService.findOne({ id });
+  async findOne(
+    @Param('id', ParseIntPipe) id: number,
+    @CaslAbility() ability,
+  ): Promise<Role> {
+    if (ability.can(Action.Read, 'role')) {
+      return this.roleService.findOne({ id });
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
-  @UseInterceptors(DurationInterceptor)
   @Post('create')
   async create(
     @Body(new ClassValidationPipe()) role: CreateRoleInput,
+    @CaslAbility() ability,
   ): Promise<Role> {
-    return this.roleService.create(role);
+    if (ability.can(Action.Create, 'role')) {
+      return this.roleService.create(role);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
-  @UseInterceptors(DurationInterceptor)
   @Put(':id')
   async updateThis(
     @Param('id', ParseIntPipe) id: number,
     @Body() role: UpdateRoleInput,
+    @CaslAbility() ability,
   ): Promise<Role> {
-    return this.roleService.update(role);
+    if (ability.can(Action.Update, 'role')) {
+      return this.roleService.update(role);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
-  @UseInterceptors(DurationInterceptor)
   @Delete(':id')
-  async delete(@Param('id', ParseIntPipe) id: number): Promise<number> {
-    return this.roleService.delete(id);
+  async delete(
+    @Param('id', ParseIntPipe) id: number,
+    @CaslAbility() ability,
+  ): Promise<number> {
+    if (ability.can(Action.Delete, 'role')) {
+      return this.roleService.delete(id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 }
